\name{generate_a_sequence}
\alias{generate_a_sequence}
\title{randomly generate a DNA coding sequence}
\description{
Given the number of codons, randomly generate a DNA coding sequence.
}
\usage{
generate_a_sequence(n_codon)
}
\arguments{
  \item{n_codon}{number of codon in the sequence}
}
\examples{
my_sequence = generate_a_sequence(20)

}
